pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Artyom", email = "artyom.groshev2017@gmail.com"},
]
dependencies = [
    "hypercorn>=0.14.3",
    "fastapi>=0.89.1",
    "pydantic>=1.10.4",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg-binary>=3.1.8",
    "python-dotenv>=0.21.1",
    "email-validator>=1.3.1",
    "alembic>=1.9.2",
    "psycopg>=3.1.8",
    "redis>=4.5.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "hypercorn --bind 0.0.0.0:8000 --reload src.main:app"
docker-dev = "docker compose -f docker-compose-dev.yml up --build"
